Kemal Akyüz CV
Hello, my name is Kemal Akyüz. I was born in Mersin in 2007. I learned to code for the first time in 2015. I started to improve myself by putting the codes I wrote on it. I did my first job in 2017. In 2018, I started my own software company by developing it myself. I started to do domestic and international projects with the company I founded. In 2022, I took part in 2 TÜBİTAK and 1 Teknofest projects as a software developer. My goal is to open my own office and serve companies. Burak Bozkurt, the founder and CEO of 51 Digital, is my biggest idol.

2017 - Baskent Tourism | Web Software - Social Media Marketing - Funnel Setup
2017 - Bursa Kayaalp Logistics | Web Software - SEO Consulting - Production & Advertising
2018 - Trendyol | Web Design - Mobile application development
2018 - Life Has Come to Our House | Web Design and Software - Social Media Marketing - Cyber ​​Security
2019 - Dietitian | Web Design and Software - Mobile application development
2020 - Akyüz News | Web Software - SEO Consulting
2022 - TUBITAK 4006 Projects | Software - Project Consulting
2022 - TUBITAK 4006 Projects | Software
2022 - TEKNOFEST EC Races | Software - Project Consulting - Social Media Management